requirements.txt
torch==2.11.0
torchvision==0.26.0
git+https://github.com/MedMNIST/MedMNIST.git
mlflow==3.10.1
streamlit==1.53.1
transformers==5.1.0
scikit-learn==1.8.0
pandas==2.3.3
matplotlib>=3.10
seaborn>=0.13
Pillow>=12.1
PyYAML>=6.0
tqdm>=4.67
